# src/generated/types.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.7.1 DO NOT EDIT.
package generated

import (
	"time"
)

// Defines values for ErrorType.
const (
	ErrorTypeCONFLICT ErrorType = "CONFLICT"

	ErrorTypeINVALIDREQUEST ErrorType = "INVALID_REQUEST"

	ErrorTypeNOTFOUND ErrorType = "NOT_FOUND"
)

// Defines values for StatusField.
const (
	StatusFieldAPPROUVED StatusField = "APPROUVED"

	StatusFieldARCHIVED StatusField = "ARCHIVED"

	StatusFieldCHANGESREQUESTED StatusField = "CHANGES_REQUESTED"

	StatusFieldDRAFT StatusField = "DRAFT"

	StatusFieldNEEDREVIEW StatusField = "NEED_REVIEW"

	StatusFieldNEW StatusField = "NEW"

	StatusFieldONREVIEW StatusField = "ON_REVIEW"

	StatusFieldREJECTED StatusField = "REJECTED"
)

// Error defines model for Error.
type Error struct {
	Message *string    `json:"message,omitempty"`
	Type    *ErrorType `json:"type,omitempty"`
}

// ErrorType defines model for Error.Type.
type ErrorType string

// full body
type Specification struct {

	// login of creator
	Author *string `json:"author,omitempty"`

	// plain text specification body json
	Body      *string    `json:"body,omitempty"`
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// comment
	Description *string `json:"description,omitempty"`

	// unique specification name, used like id
	Name *string `json:"name,omitempty"`

	// status uses for collaboration processes
	Status    *StatusField `json:"status,omitempty"`
	UpdatedAt *time.Time   `json:"updated_at,omitempty"`
}

// fields set for create specification
type SpecificationCreateBody struct {

	// comment
	Description *string `json:"description,omitempty"`

	// unique specification name, used like id
	Name *string `json:"name,omitempty"`
}

// Specifications defines model for Specifications.
type Specifications []Specification

// status uses for collaboration processes
type StatusField string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse Error

// CreateSpecificationJSONBody defines parameters for CreateSpecification.
type CreateSpecificationJSONBody SpecificationCreateBody

// RetrieveSpecificationsParams defines parameters for RetrieveSpecifications.
type RetrieveSpecificationsParams struct {

	// filter by status
	Status *string `json:"status,omitempty"`

	// filter by author
	Author *string `json:"author,omitempty"`
}

// CreateSpecificationJSONRequestBody defines body for CreateSpecification for application/json ContentType.
type CreateSpecificationJSONRequestBody CreateSpecificationJSONBody